+++ cli/src/commands/import/csv.ts
@@ -40,6 +40,20 @@ export default class ImportCSV extends BaseCommand {
     }),
     'no-column-name-normalization': Flags.boolean({
       description: 'Avoid changing column names in a normalized way'
+    }),
+    'batch-size': Flags.integer({
+      description: 'Batch size to process and upload records'
+    }),
+    'max-rows': Flags.integer({
+      description: 'Maximum number of rows to process'
+    }),
+    delimiter: Flags.string({
+      description: 'Delimiters to use for splitting CSV data',
+      multiple: true
+    }),
+    'null-value': Flags.string({
+      description: 'Value to use for null values',
+      multiple: true
     })
   };
 
@@ -54,7 +68,11 @@ export default class ImportCSV extends BaseCommand {
       columns,
       'no-header': noHeader,
       create,
-      'no-column-name-normalization': ignoreColumnNormalization
+      'no-column-name-normalization': ignoreColumnNormalization,
+      'batch-size': batchSize,
+      'max-rows': maxRows,
+      delimiter,
+      'null-value': nullValue
     } = flags;
 
     const { workspace, database, branch } = await this.getParsedDatabaseURLWithBranch(flags.db, flags.branch);
@@ -72,6 +90,10 @@ export default class ImportCSV extends BaseCommand {
       types: splitCommas(types),
       columns: splitCommas(columns),
       noheader: Boolean(noHeader),
+      batchSize,
+      maxRows,
+      delimiter,
+      nullValue,
       ignoreColumnNormalization,
       shouldContinue: async (compare) => {
         return Boolean(await this.shouldContinue(compare, table, create));