+++ cli/src/schema.ts
@@ -5,6 +5,7 @@ export type Column = {
   name: string;
   type: 'bool' | 'int' | 'float' | 'string' | 'text' | 'email' | 'multiple' | 'link' | 'object' | 'datetime';
   unique?: boolean;
+  notNull?: boolean;
   description?: string;
   link?: {
     table: string;
@@ -17,6 +18,7 @@ export const columnSchema: z.ZodSchema<Column> = z.lazy(() =>
     name: z.string(),
     type: z.enum(['bool', 'int', 'float', 'string', 'text', 'email', 'multiple', 'link', 'object', 'datetime']),
     unique: z.boolean().optional(),
+    notNull: z.boolean().optional(),
     description: z.string().optional(),
     link: z
       .object({