+++ test/integration/create.test.ts
@@ -1,56 +1,41 @@
-import fetch from 'cross-fetch';
-import dotenv from 'dotenv';
-import { join } from 'path';
-import { afterAll, beforeAll, describe, expect, test } from 'vitest';
-import { XataApiClient } from '../../packages/client/src';
+import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, test } from 'vitest';
 import { XataClient } from '../../packages/codegen/example/xata';
-import { teamColumns, userColumns } from '../mock_data';
+import { setUpTestEnvironment, TestEnvironmentResult } from '../utils/setup';
 
-// Get environment variables before reading them
-dotenv.config({ path: join(process.cwd(), '.env') });
+let xata: XataClient;
+let hooks: TestEnvironmentResult['hooks'];
 
-let client: XataClient;
-let databaseName: string;
+beforeAll(async (ctx) => {
+  const result = await setUpTestEnvironment('create');
 
-const apiKey = process.env.XATA_API_KEY ?? '';
-const workspace = process.env.XATA_WORKSPACE ?? '';
-if (workspace === '') throw new Error('XATA_WORKSPACE environment variable is not set');
+  xata = result.client;
+  hooks = result.hooks;
 
-const api = new XataApiClient({ apiKey, fetch });
-
-beforeAll(async () => {
-  const id = Math.round(Math.random() * 100000);
-
-  const database = await api.databases.createDatabase(workspace, `sdk-integration-test-create-${id}`);
-  databaseName = database.databaseName;
-
-  client = new XataClient({
-    databaseURL: `https://${workspace}.xata.sh/db/${database.databaseName}`,
-    branch: 'main',
-    apiKey: process.env.XATA_API_KEY || '',
-    fetch
-  });
-
-  await api.tables.createTable(workspace, databaseName, 'main', 'teams');
-  await api.tables.createTable(workspace, databaseName, 'main', 'users');
-  await api.tables.setTableSchema(workspace, databaseName, 'main', 'teams', { columns: teamColumns });
-  await api.tables.setTableSchema(workspace, databaseName, 'main', 'users', { columns: userColumns });
+  return hooks.beforeAll(ctx);
 });
 
-afterAll(async () => {
-  await api.databases.deleteDatabase(workspace, databaseName);
+afterAll(async (ctx) => {
+  await hooks.afterAll(ctx);
+});
+
+beforeEach(async (ctx) => {
+  await hooks.beforeEach(ctx);
+});
+
+afterEach(async (ctx) => {
+  await hooks.afterEach(ctx);
 });
 
 describe('record creation', () => {
   test('create single team without id', async () => {
-    const team = await client.db.teams.create({ name: 'Team ships' });
+    const team = await xata.db.teams.create({ name: 'Team ships' });
 
     expect(team.id).toBeDefined();
     expect(team.name).toBe('Team ships');
   });
 
   test('create multiple teams without ids', async () => {
-    const teams = await client.db.teams.create([{ name: 'Team cars' }, { name: 'Team planes' }]);
+    const teams = await xata.db.teams.create([{ name: 'Team cars' }, { name: 'Team planes' }]);
 
     expect(teams).toHaveLength(2);
     expect(teams[0].id).toBeDefined();
@@ -62,24 +47,25 @@ describe('record creation', () => {
   });
 
   test('create user with id', async () => {
-    const user = await client.db.users.create('a-unique-record-john-4', {
+    const user = await xata.db.users.create('a-unique-record-john-4', {
       full_name: 'John Doe 4',
       email: 'john4@doe.com'
     });
 
-    const apiUser = await client.db.users.filter({ id: user.id }).getFirst();
+    const apiUser = await xata.db.users.filter({ id: user.id }).getFirst();
     if (!apiUser) throw new Error('No user found');
 
     expect(user.id).toBe('a-unique-record-john-4');
     expect(user.read).toBeDefined();
     expect(user.full_name).toBe('John Doe 4');
+    expect(user.full_name.startsWith('John')).toBe(true);
 
     expect(user.id).toBe(apiUser.id);
     expect(user.full_name).toBe(apiUser.full_name);
     expect(user.email).toBe(apiUser.email);
 
     expect(
-      client.db.users.create('a-unique-record-john-4', {
+      xata.db.users.create('a-unique-record-john-4', {
         full_name: 'John Doe 5',
         email: 'john5@doe.com'
       })
@@ -87,13 +73,13 @@ describe('record creation', () => {
   });
 
   test('create user with inlined id', async () => {
-    const user = await client.db.users.create({
+    const user = await xata.db.users.create({
       id: 'a-unique-record-john-5',
       full_name: 'John Doe 5',
       email: 'john5@doe.com'
     });
 
-    const apiUser = await client.db.users.filter({ id: user.id }).getFirst();
+    const apiUser = await xata.db.users.filter({ id: user.id }).getFirst();
     if (!apiUser) throw new Error('No user found');
 
     expect(user.id).toBe('a-unique-record-john-5');
@@ -107,7 +93,7 @@ describe('record creation', () => {
 
   test('create user with empty id is not allowed', async () => {
     expect(
-      client.db.users.create('', {
+      xata.db.users.create('', {
         full_name: 'John Doe 3',
         email: 'john3@doe.com'
       })
@@ -116,7 +102,7 @@ describe('record creation', () => {
 
   test('create user with empty inline id is not allowed', async () => {
     expect(
-      client.db.users.create({
+      xata.db.users.create({
         id: '',
         full_name: 'John Doe 3',
         email: 'john3@doe.com'
@@ -127,7 +113,7 @@ describe('record creation', () => {
   test('create user with falsy id is not allowed', async () => {
     expect(
       //@ts-expect-error
-      client.db.users.create(null, {
+      xata.db.users.create(null, {
         full_name: 'John Doe 3',
         email: 'john3@doe.com'
       })
@@ -135,12 +121,12 @@ describe('record creation', () => {
   });
 
   test("create multiple with empty array doesn't create anything", async () => {
-    const teams = await client.db.teams.create([]);
+    const teams = await xata.db.teams.create([]);
     expect(teams).toHaveLength(0);
   });
 
   test('create multiple some with id and others without id', async () => {
-    const teams = await client.db.teams.create([{ id: 'team_cars', name: 'Team cars' }, { name: 'Team planes' }]);
+    const teams = await xata.db.teams.create([{ id: 'team_cars', name: 'Team cars' }, { name: 'Team planes' }]);
 
     expect(teams).toHaveLength(2);
     expect(teams[0].id).toBe('team_cars');
@@ -152,10 +138,7 @@ describe('record creation', () => {
   });
 
   test('create multiple with returning columns', async () => {
-    const teams = await client.db.teams.create(
-      [{ name: 'Team cars' }, { name: 'Team planes', labels: ['foo'] }],
-      ['id']
-    );
+    const teams = await xata.db.teams.create([{ name: 'Team cars' }, { name: 'Team planes', labels: ['foo'] }], ['id']);
 
     expect(teams).toHaveLength(2);
     expect(teams[0].id).toBeDefined();
@@ -179,7 +162,7 @@ describe('record creation', () => {
   });
 
   test('create single with returning columns', async () => {
-    const team = await client.db.teams.create({ name: 'Team cars' }, ['id']);
+    const team = await xata.db.teams.create({ name: 'Team cars' }, ['id']);
 
     expect(team).toBeDefined();
     expect(team.id).toBeDefined();
@@ -190,5 +173,22 @@ describe('record creation', () => {
     const team1 = await team.read();
     expect(team1?.id).toBe(team.id);
     expect(team1?.name).toBe('Team cars');
+  });
+
+  test('create single with unique email', async () => {
+    const data = { full_name: 'John Doe 3', email: 'unique@example.com' };
+    const user = await xata.db.users.create(data);
+
+    expect(user.id).toBeDefined();
+    expect(user.read).toBeDefined();
+    expect(user.full_name).toBe(data.full_name);
+    expect(user.email).toBe(data.email);
+
+    await expect(xata.db.users.create(data)).rejects.toThrowError();
+  });
+
+  test('create single with notNull column', async () => {
+    // @ts-expect-error
+    await expect(xata.db.users.create({})).rejects.toThrowError();
   });
 });